requirements.txt
flask
flask-cors
twilio
python-dotenv